singleton: Become process group leader so when the singleton is killed, COMMAND is properly cleaned up as well
When COMMAND becomes process group leader (e.g. onchange does this), it is decoupled from the current process, so when the singleton is killed, the COMMAND will linger on, and we don't want that; the lifetime of COMMAND should be coupled to the lifetime of singleton (to be able to enforce the singleton-ness).
Therefore, even though singleton itself doesn't launch background commands, become process group leader and do the killing. The exported PGID indicates to COMMAND that it itself does not need to become leader.