runAll: Drop isNeedEval, use all arguments as one SIMPLECOMMAND, support {} marker
It looks like this was derived from a related command, and not well thought through :-(
The command-line argument parsing took each SIMPLECOMMAND as a separate COMMAND, which rarely makes sense, and goes against conventions. Rather, take all SIMPLECOMMAND [...] as _one_ command.
For non-prompt, the "$c" execution will only work for a standalone SIMPLECOMMAND. As the prompt branch also just switches the quoter, we can completely get rid of the isNeedEval flag and always quote; that will also support SIMPLECOMMAND with additional arguments. Plus, we don't need the pre-parsing at all!
After that, we can simply support an argument marker ({}); for the non-prompt branch in the canonical way, with prompt by explicitly testing the command for {} and distinguishing both cases.