runAll: FIX: Delegate to runWithPrompt
The usages got messed up and are outdated (e.g. --exit-unless was missing). Still duplicate the command-line argument parsing (no good solution there), but at least delegate to the original command for its short and long usages. (That reference to the command in the long usage is ugly, and also got messed up.)