runAll: FIX: Actually use runWith if any runWithPrompt args are given (cp. inkarkat/shell-tools@c3b95ab)
The default $runWithCommand never saw action because $with is only changed by --with-prompt. Do away with a separate $runWithCommand and use $with for the selection as well.
Robustness: Handle with=header via fall-through and assert on unknown $what values.

(This relates to commit inkarkat/shell-tools@c3b95abe6d20db779acd5fd6077993880cc6b5b0)