Avoid "bash: cannot set terminal process group (...): Inappropriate ioctl for device bash: no job control in this shell"
When an alias is invoked as COMMAND though addAliasSupport, that shell disables job control via set +m, causing setsid to fail with:
> bash: cannot set terminal process group (16776): Inappropriate ioctl for device
> bash: no job control in this shell
In these cases, the job control is owned by a parent process (in my example of $ withLessa cat ~/.vimrc, it's the bash started by tmux-less).
I think it should be fine to leave the supervisory role with that process (at least with tmux-less, its lifetime is closely bound to ensurePrompting and the COMMAND it's running, anyway).
Just detect this and use the existing PGID, as if we were given the process group.