ENH: Optimization: Compare COMMAND's runtime against threshold and skip pager watching if it's mostly long-running
The previous optimization took care of commands that themselves directly employ the pager, so no external prompting is necessary; the user simply quits the internal pager and that concludes the command. But there are other commands like text editors, newsreader, or process viewer that are started once and then remain running (e.g. in a separate Tmux window). These don't use a pager, but like it need to be explicitly quit by the user. The command's overall runtime can be used as a heuristic. It may be wrong; there are batch commands that simply take a long time (like encoding a video with ffmpeg) and then quit without confirmation. Without an explicit ID, the whole command it taken, so it's less likely that the same long-running command is repeated again and again. We also postpone the verdict until 10 invocations have been seen. To avoid that a few erroneous invocations (where the command immedately quits with an error) disrupt the evaluation, we don't require a perfect long-running behavior (as with the internal pager use), but allow for 10% short runs.
As with the internal pager, tallying stops once the command is deemed self-sufficient with regard to prompting. After all, the database update takes time, and this is about avoiding the overhead of polling for pager use.

System-info: Armbian 24.11.1 noble, aarch64
Platform-info: bash 5.2.21(1)-release